chore: transfer program-logic to ProgramView
