+++ pages/program.tsx
@@ -5,128 +5,29 @@ import { program_QueryResponse } from '../__generated__/program_Query.graphql';
 import PageView from '../components/PageView';
 import ProgramView from '../components/ProgramView';
 import Flex, { FlexItem } from 'styled-flex-component';
-import { groupBy, sortBy } from 'lodash';
-import dayjs from 'dayjs';
-import styled from 'styled-components';
-import ReactMarkdown from 'react-markdown';
-
-import {
-  Heading,
-  Blockquote,
-  ThematicBreak,
-  MarkdownList,
-  Paragraph,
-} from '../components/MarkdownRenderer';
-import Link from 'next/link';
-import { ArrayElement, eventTime } from '../components/Stands/StandCard';
-
-type EventType = ArrayElement<NonNullable<program_QueryResponse['events']>>;
-
-interface LocationLinkProps {
-  event: EventType;
-  stands: program_QueryResponse['stands'];
-}
-
-const LocationLink = ({ event, stands }: LocationLinkProps): JSX.Element => {
-  let href;
-
-  const standSlug =
-    stands &&
-    stands.find((stand) => stand && stand.company.id === event.company?.id)
-      ?.slug;
-
-  // FIXME: This should be implemented as types
-  switch (event.location.toLowerCase()) {
-    case 'forsiden':
-      href = '/stands';
-      break;
-    case 'standen':
-      href = standSlug ? `stands/${standSlug}` : null;
-      break;
-    default:
-      href = null;
-      break;
-  }
-
-  return href ? (
-    <InfoElement>
-      <Link href={href}>
-        <HostingCompanyLink>{`📍 ${event.location}`}</HostingCompanyLink>
-      </Link>
-    </InfoElement>
-  ) : (
-    <InfoElement>
-      <HostingCompanyNoLink>{`📍 ${event.location}`}</HostingCompanyNoLink>
-    </InfoElement>
-  );
-};
 
 const Index = ({
   error,
   props,
-}: WithDataAndLayoutProps<program_QueryResponse>): JSX.Element => {
-  const groupedEvents =
-    props.events && groupBy(sortBy(props.events, 'timeStart'), 'date');
-  const sortedKeys = props.events && sortBy(Object.keys(groupedEvents || {}));
+}: WithDataAndLayoutProps<program_QueryResponse>): JSX.Element => (
+  <>
+    {props.programPage && <PageView hideContent page={props.programPage} />}
 
-  return (
-    <>
-      {props.programPage && <PageView hideContent page={props.programPage} />}
+    {props.programPage && (
+      <PageView hideTitle hideDate page={props.programPage} />
+    )}
 
-      {props.programPage && (
-        <PageView hideTitle hideDate page={props.programPage} />
-      )}
-
-      {/* {props.events ? (
-        <ProgramView events={props.events} /> */}
-      {sortedKeys && groupedEvents ? (
-        <CenterFlex>
-          {sortedKeys.map((k) => (
-            <GroupedDateEvent key={k}>
-              <DateTitle>
-                {dayjs(k).format('dddd DD.MM').toUpperCase()}
-              </DateTitle>
-              <GroupedEvent>
-                {groupedEvents[k].map((event) => (
-                  <EventInfo key={event.id}>
-                    <Title>{event.title}</Title>
-                    <EventTimePlaceInfo>
-                      <InfoElement>{`🕐 ${
-                        eventTime(event).timeRange
-                      }`}</InfoElement>
-                      <LocationLink event={event} stands={props.stands} />
-                      {event.company && (
-                        <HostingCompanyNoLink>{`🏢 ${event.company?.name}`}</HostingCompanyNoLink>
-                      )}
-                    </EventTimePlaceInfo>
-                    <br />
-                    <ReactMarkdown
-                      source={event.description}
-                      escapeHtml={false}
-                      renderers={{
-                        heading: Heading,
-                        blockquote: Blockquote,
-                        thematicBreak: ThematicBreak,
-                        list: MarkdownList,
-                        paragraph: Paragraph,
-                      }}
-                    />
-                  </EventInfo>
-                ))}
-              </GroupedEvent>
-            </GroupedDateEvent>
-          ))}
-        </CenterFlex>
-      ) : (
-        <Flex>
-          <FlexItem>
-            <h1>Programmet er tomt</h1>
-          </FlexItem>
-        </Flex>
-      )}
-    </>
-  );
-};
+    {props.events ? (
+      <ProgramView events={props.events} stands={props.stands} />
+    ) : (
+      <Flex>
+        <FlexItem>
+          <h1>Programmet er tomt</h1>
+        </FlexItem>
+      </Flex>
+    )}
+  </>
+);
 
 export default withDataAndLayout(Index, {
   query: graphql`
@@ -167,94 +68,3 @@ export default withDataAndLayout(Index, {
     metadata: props && props.programPage,
   }),
 });
-
-const Title = styled.h2`
-  position: relative;
-  margin: 0px;
-  &::after {
-    content: ' ';
-    position: absolute;
-    background-color: #f5f5f5;
-    border: 3.5px solid #0d7bb4;
-    width: 15px;
-    height: 15px;
-    bottom: 9px;
-    left: -63.2px;
-    box-shadow: -2px 0px 5px 2px rgba(0, 0, 0, 0.2);
-    border-radius: 50%;
-  }
-  @media only screen and (max-width: 991px) {
-    font-size: 1.7em;
-
-    &::after {
-      bottom: 2px;
-    }
-  }
-`;
-
-const DateTitle = styled.h1`
-  margin: 0px 0px 5px 0px;
-  font-weight: 300;
-`;
-
-const GroupedDateEvent = styled.div`
-  display: flex;
-  margin: 0px 20px 40px;
-  flex-direction: column;
-  flex: 1 1 500px;
-  @media only screen and (max-width: 991px) {
-    flex-direction: column;
-    padding: 0px 0px;
-    margin: 0px 20px 10px;
-  }
-`;
-
-const CenterFlex = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  margin-top: 50px;
-  @media only screen and (max-width: 991px) {
-    padding: 0px;
-    margin: 0px;
-  }
-`;
-
-const GroupedEvent = styled.div`
-  display: flex;
-  align-items: flex-start;
-  flex-direction: column;
-  box-shadow: 0px 0px 5px 3px rgba(0, 0, 0, 0.2);
-  border-radius: 0px 10px 10px 0px;
-  padding: 0px 50px 30px 50px;
-  border-left: 4px solid #0d7bb4;
-  @media only screen and (max-width: 991px) {
-    margin-bottom: 50px;
-  }
-`;
-
-const EventTimePlaceInfo = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-`;
-
-const InfoElement = styled.div`
-  margin-right: 15px;
-`;
-
-const HostingCompanyLink = styled.a`
-  font-weight: 500;
-  cursor: pointer;
-`;
-
-const HostingCompanyNoLink = styled.span`
-  font-weight: 400;
-`;
-
-const EventInfo = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-top: 30px;
-`;
